+++ src/main/java/com/hhplus/concert_ticketing/infra/concert/JpaPerformanceRepository.java
@@ -1,4 +1,4 @@
-package com.hhplus.concert_ticketing.domain.concert;
+package com.hhplus.concert_ticketing.infra.concert;
 
 import com.hhplus.concert_ticketing.domain.concert.ConcertStatus;
 import com.hhplus.concert_ticketing.domain.concert.ConcertPerformance;
@@ -7,4 +7,6 @@ import java.util.List;
 
 public interface JpaPerformanceRepository {
     List<ConcertPerformance> findByConcertIdAndAvailableSeatsGreaterThanAndStatusIn(Long concertId, int availableSeat, ConcertStatus status);
+
+    ConcertPerformance findById(Long performanceId);
 }
